Adds CLA workflow via github actions
See https://github.com/cla-assistant/github-action

Instead of manually having to manage this, we can have a
bot do it for us.

This will mean that people who contribute will agree and then
this will store this in a file in this repository.

Note, if we don't want Stitch Fix people to be hit by this,
we just need to add them to the allow list.