Adds ability to change augment Hamilton execution & return type
Hamilton was hardcoded to return a pandas dataframe. This change
introduces a module `base` that houses:

1. an interface object `ResultMixin` with a static function called "build_result".
2. an interface object `HamiltonGraphAdapter` that houses the required interface to augment execution more broadly.

`ResultMixin`: The idea is that we'll use this class as the interface to describe the shape required to return different results.
I then went ahead and implemented DictResult, PandasDataFrameResult, and NumpyMatrixResult.

`HamiltonGraphAdapter`: paves the way for distributed computation & specifying different return types. Since it allows us to augment and wrap Hamilton
functions as we're walking the graph. All GraphAdapters will therefore inherit and override the functions in `HamiltonGraphAdapter`.
I decided to have it extend ResultMixin to simplify things. In some cases the HamiltonGraphAdapter will have its own need
to stitch together a result, in other cases it will just delegate to a mixin, or via delegation.

Why those two "type" functions in `HamiltonGraphAdapter`? We do a static time compile check for equivalence when building a graph, and
one when we're inspecting inputs. We need to delegate this check to the adapter in case there are type mismatches
because some might be expected (e.g. dask can mimic pandas). I don't necessarily like wiring through the
adapter all the way through, but it works; I think we'll be able to change things
easily since basically all the code in graph.py should be viewed as internal code.

Otherwise all these changes should be backwards compatible -- passing in a HamiltonGraphAdapter is optional. The
`SimplePythonDataFrameHamiltonGraphAdapter`, which mirrors the current Hamilton behavior is the default if none
is provided.

Comment on FunctionGraph: I expect to further augment FunctionGraph to tease out certain parts of it to enable
a broader array of use cases. Otherwise I think we're pretty fixed on "black-box" reuse, so getting reuse of the
FunctionGraph in different context by passing in a object of a set interface, versus using inheritance.