Adds integration test for h_spark
This mirrors the h_dask and h_ray tests.

Adds this to run in circleci.

Note regarding circleci: since we need Java we
have to change the image. The image has python
but not pip. So I need to do some apt-get things,
and use aptitude to install things so that python
is in a working state on the image. It does complain
about bdist not installing with pyspark, however
that doesn't seem to break actually running the code.