Refactors visualization and fixes #41
Visualization was an afterthought we added quickly.
This change takes in feedback from #45 and makes
it more amenable for actual use.

We therefore expose new functions, and deprecate
the `visualize` parameter  in `driver.execute()`.

The new functions enable the user to pass in a dictionary that will
be used as key word arguments (kwargs) to the
render function -- thereby giving lots of control on
how things are visualized.

Along the way I also refactored the cycle detection
using networkx. This again, was a quick hack,
now it's a standalone implementation. We do break graph
 behavior, but not driver behavior with this refactor. I
believe that's fine, since we don't want people to
depend on FunctionGraph directly just yet.

Adds unit tests for these changes.