Adds extract_fields decorator for operating over dicts
The API to use it looks like this:

```python

@function_modifiers.extract_fields(
    {'X_train': np.ndarray, 'X_test': np.ndarray, 'y_train': np.ndarray, 'y_test': np.ndarray})
def train_test_split_func( ... ,  ... ) -> dict
```

I decided to go with a straight dict of `field_name` to `field_type` because that seemed
the simplest thing to define. Note, we use the documentation for the original function,
rather than enabling individual doc strings for the types. I think this suffices for now.

To support TypedDict, I didn't want to have to import typing_extensions to handle
it. Also you can't inline define a TypedDict class, so it would be more verbose which
is less that ideal.

We can always add TypedDict support later. Also punted on `Tuple` support -- that
might be another decorator...