Avoids double execution of nodes
Previously we were not checking if a node was in "computed". Now we are,
short-circuiting the computation of any dependencies. Note that variables
specified in overrides are not *yet* checked for conflicts with computed.
This is something we will fix, but for now its undefined behavior.

A test erroneously relied on that. This changes the assumptions in the test.