Comments out overrides tests with a cyclic graph
We don't officially want to support this -- but this is possible.
i.e. given a graph that is cyclic, if you can make it acyclic, then
computation will work -- you'd do this via overrides.

In future, we may change how we do execute, and this could
become burdensome. So punting on making this something
we truly support; if it's not tested, it ain't supported.