Ensures all functions have module tag defined and propagated
We should add default tags to all functions when appropriate. I think
one of the appropriate things to do, is to include the module that the
function is in. This will be useful for people who create a DAG from
multiple modules and want some ability to find/query/subset functions.

I didn't add any other default ones, because I think we can think
about what they should be a little more. E.g. does it make sense for
extract_* to mention that the node created was done via the decorator? To
me it doesn't seem like we need that. When we add data quality, I think
there will be more of a use case there to tag "nodes" that are what data
quality added for instance.

Note: we want to copy the tags where appropriate, because we don't want
a later stage to modify their tag dict and have it propagated to the others by
accident.