Exposes passing kwargs to graphviz object
This should be a backwards compatible API change for everything.

To be able to control, for instance the ratio, of the produced image,
it's not a render time argument. Instead, it's a property of the graph.
To be able to manipulate it, we need to wire through passing in attributes.
e.g. https://graphviz.org/doc/info/attrs.html shows what could be set.

These attributes are either graph, node, or edge attributes.
The kwargs I'm specifically trying to expose are:
```
graph_attr – Mapping of (attribute, value) pairs for the graph.
node_attr – Mapping of (attribute, value) pairs set for all nodes.
edge_attr – Mapping of (attribute, value) pairs set for all edges.
```
The above was taken from https://graphviz.readthedocs.io/en/stable/api.html?highlight=ratio#digraph.

Note, this will  enable anything else to be also passed to the graph, which I
think is a useful thing too. Since that will allow someone who knows graphviz
to really customize it I think.