Adds docs warning ad hoc module use with Ray, Dask, Spark
The ability to create temporary modules that house functions that
was just added does not work for scaling onto Ray, Dask, or Pandas on
Spark. That's because we add it manually to sys.modules, but these frameworks
pickle the function, and then expect an actual module to exist on the other side
to unpickle the function.