Adds DataInValuesValidatorP* validators
PandasSeries and Primitives are added.

This enables one to do:

```python
@check_output(values_in=[ LIST,  OF, VALUES ], ...)
```
Which will check what the function outputs and validate that it
is within one of the values provided.

Currently the primitives operates over numbers and strings.
I punted on lists and dictionaries -- they should probably be
different validator classes.