Adds data quality example
This data quality example is based on the example
we provided with the outerbounds(metaflow) folks.

It's purpose is to show how one might apply data quality.
It also shows how to use the same code and make it
work for dask*, ray, spark.

Ray - everything seems to work as intended.

Spark - had to change how the data & some features
are computed.

Dask* - had to change how the data & some features
are computed.  CAVEAT: right now the validators don't work properly
for dask data types.  That is because either (1) it's a future object, or (2)
we use pandas series syntax, when instead we should
use the dask specific syntax. In short - DEFAULT DQ DOES NOT WORK WITH
DASK DATA TYPES. BUT it DOES WORK if you're just using spark
for multi-processing, and not using dask data types.

So we need to think how we might change/inject the validator implementation
based on say a graph adapter or something... otherwise this forces one
to really stick to one data type or another, i.e. pandas or dask dataframe.

Documentation should hopefully be enough to document what is doing on.
The only TODO is to create an analogous example using Pandera -- my hope
is that it will handle dask datatypes...