Enables primitive validators to operate over dask datatypes
Using the same trick as we employed before, we can simply
compute a result for the scalar primitive validators to be a valid
value to compare against.

Without this, things break because we're trying to compare
a dask data type thing.

Note: we could do a similar strategy for the Pandas Series validators,
however we'd need to do something akin to what pandera does
under the hood with `map_partitions` over the dask like object.
I vote to push people to use pandera if they're using dask data types.