Adds `parametrized_full` function modifier
This attempts to unit all the parametrized flavors of
function_modifiers. Specifically, this enables one to specify two source
of inputs:

1. literal
2. upstream

Replacing the function's arguments, as well as direct assignment to a
node name. Note that, currently, docstrngs are provided as part of the
decorator. We will be allowing this to use parametrized docstrings.

This intends to replace all types of parametrized decorators using delegation.