Fixes DAG construction slowness
This is an issue with check_output. We were using sum(lists) which does
a concat operation for each list. This is fine at a small scale, but at
a larger scale this is an O(n^2) operation. We can still optimize this
significantly but this fixes the underlying issue in #168.

Note that this does *not* fix the issues around ray yet, this is just
the underlying piece. With Ray, the hard part is validating everything
which is inherently slow. Worth looking into why parallelism didn't help
though.