Simplifies async driver API
This removes the ability to add coroutines as inputs. It was flaky,
and suboptimal. If we want to do that we can do it later, but for
now pure data has to be passed in. Note that this was never released
(and never worked), and is experimental, so we can happily remove this
without a major bump.