Tweaks warning message from pandas index check
If using a dask dataframe or series the index type checker
will not capture it because it only checks for pandas types.
So adding note to ignore if using DASK.

We will have to figure out how to support this -- likely
with a specific result builder for Dask that takes this into
account, or something like that.