Implements parameterize_extract_columns + a dataframe API
This provides a convenience API for
https://github.com/stitchfix/hamilton/issues/196. The idea is that
people want to do a parameterization and extract columns operator at
once -- this should be easy. The cool thing is that this just uses the
parameterize and extract APIs. It also has a from_df() function to allow
for passing in a dataframe to be more concise.