Improves async example to use proper function call
We do not want to show case `raw_execute` anywhere.
People should be using `execute()` exclusively. So
fixing this example to show how to get a dictionary back,
rather than using `raw_execute`.

Also added documentation to show you can instantiate
it in a function, or outside in the module.