Introduces *_extensions to extend functionality in hamilton
So we had a problem in the decorators (notably extract_columns)
that assumed pandas. We need to abstract this out so that other
dataframe libraries could be easily used with the decorator.

This is a commit that introduces a few changes to do that:

(1) We decided on a plugins module that houses two types of files:
*_extensions.py, and *_implementations.py. The *_extensions includes
code that the decorators and hamilton framework depends on.
The *_implementations will house specific implementations related to
that dataframe library for example. Note we haven't created any _implementations.py
files in this commit -- will do so in later ones.

(2) We decide that there are several functions that need to live in _extensions.py
that enable the hamilton framework to delegate/find the right thing to use.
This could be an object -- that one has to implement. But because it's already
in the module -- and we want things to be registered, I think this abstraction right
now is simpler -- the housing unit is the module that the functions are in -- don't need
to put them into an object right now.

(3) I decided trying to import the modules and having them determine how to error
would be the best way to determine if the rest of the module import should proceed.
We then check the modules have correctly expose a few things to check the interface
is consistent between all of them.