Refactors and creates a registry module
So instead of putting it in base, I decided to move
the extension registering logic to the `registry` module.
This module is not allowed to depend on anything else.

Then to trigger the extension loading, I think `hamilton.base` is
the right place to do it. Which calls the function in registry to try to
import the extensions.

There is no circular dependency with how `registry` loads
the extension modules, because it's wrapped in a function
which does not trigger until after loading the module. Then
when the extensions load, they import registry which doesn't
import anything so things are good.