Moves registry initialization to function_modifiers/base.py
Ray serialization is the cause of this move. Which I guess
is for good reason, as decorators are the only thing that
need the registry right now. So probably makes more sense to
initialize here than in hamilton/base.py anyway.

If there are future issues, we can just try initializing in both
places -- I tested that works -- but removed it from hamilton/base.py
for now.