Filters out certain tests for python3.6
We're migrating off of supporting it, its basically at EOL. So, these
ones were tricky to get to work, and aren't worth messing with the code.
To be clear, there will be no loss of functionality, just less testing.
The features that these tests test for were previously broken for all
python versions and are now just broken for 3.6.

Only caveat is that the end-to-end tests test for a few more things, but
I'm not worried about it as we're migrating off of 3.6 anyway:

https://github.com/stitchfix/hamilton/issues/267