Adds m5 forecasting kaggle example
I pretty much verbatim copied the code from [this notebook](https://www.kaggle.com/code/ratan123/m5-forecasting-lightgbm-with-timeseries-splits).
It does have a few odd quirks in it -- so if I understood what was going on I removed it, else I left it in. E.g. `test_2` is not used...

This is an example that shows:

1. loading data
2. creating a base set.
3. adding more features.
4. creating a training set.
5. fitting model & predicting with it.
6. saving the result.

There are a few other ways one could write some of this code.
But as a first pass this seems fine. E.g. we could split up the model
training function a bit more. Or we could use the parameterize decorator.

No unit tests have been added, or data quality checks.