Fixes minor bug with default arguments
There was a bug where we would consider nodes that
were not part of the execution graph to see if all the inputs
to a node were acceptable.

E.g. if two functions required the same input, but one had default arguments,
and the function without a default was not on the execution path, we'd error.
Why? Because we'd naively look at all the node's dependents to see if one was
required, rather than restricting to the set of nodes required for execution.

Adds to unit test; aside some of our driver code could be made easier to unit test...