Adds ability to pass in a callable to get the function name from for execute
I found myself wanting to click directly to outputs being created to inspect them.
So to short cut that, directly referencing them is a way I can do that and have
my IDE jump there when navigating code.

The downside is that this solution doesn't handle functions annotated with decorators
that create nodes, e.g. @parameterize*, or @config* . I think that's okay, because
it's probably clearer to have a "string" name here to indicate that something is
changing/being produced at DAG creation time, and thus a reference to an
exact function is likely a misnomer...

Adds unit test for this.