Adds code to make concatenating dataframes in the result builder default
Why this change? because if you are doing indicator variable stuff,
maybe you don't want to be explicit about them, instead you want
to concatentate the columns created without knowing about them. Seems
like a reasonable pattern on first glance.

Otherwise this isn't pretty code. Still thinking it through, but essentially:

1. want to handle column concat of dataframes and series.
2. problem is that pd.DataFrame({ dict }) behavior essentially does
an outer join on indexes if present, and handles scalars.
3. also we need to preserve "column" i.e. output order.
4. So question -- is this doing too much for the normal case
of not having a dataframe in the output? Should I structure
the code differently?

Otherwise need to add unit tests for this new function:
 - like ensuring order is preserved.
 -