Simplifies logic to handle dataframe column concatenation
This code is simpler because it delegates to pandas on how
to construct everything -- all it does is that it flattens
the dataframe, and then ensures there's no duplicate column
definitions. I added tests for this for this case.

For the future, we probably want some flexibility with respect
to "joins" and how things are stitched together since you could
envisage some use cases where people want to inner join between
things --- but until there's a common pattern/need, I think
we should just keep thinking about what should be first class,
and what we get people to do themselves.