Tweaks logging when concatenating Dataframes
We should push for a convention to help make it easy to debug
those columns and find the code if there is an issue. I don't hard code the
convention here because I don't want to preclude things. Otherwise an annoying
warning is probably good for now -- and could be removed/moved to
debug later.